refactor(app): split app orchestration hooks
